changed ParameterChange to not implement IVectorBlock since it caused several problems related to vector block count.
Added an ILayerCommand interface at the top of the inheritance chain that covers vector blocks as well as other possible commands. Left the VectorBlocks property inside ILayer for access to the "real" vector blocks, the downside is that any such access causes an enumeration now which is an unnecessary performance cost.